api/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
PyYAML
passlib[bcrypt]